+++ effects/volumeeffect.cpp
@@ -37,7 +37,7 @@ void VolumeEffect::load(QXmlStreamReader* stream) {
         stream->readNext();
         if (stream->isStartElement() && stream->name() == "volume") {
             stream->readNext();
-            volume_val->set_value(stream->text().toInt());
+            volume_val->set_value(stream->text().toFloat());
         }
     }
 }
